requirements.txt
Django==1.9
django-debug-toolbar
django-extensions
Pygments
django-email-templates
python-ldap
MySQL-python
django-tools
unicodecsv

